Fixing some major Daylight Savings Issues
Now we normalize dates going in and out of the timeline control that
were getting messed up because of daylight savings. This is still not
'correct'. The correct behaviour would be to have another hour on the
timeline when the time changes. Events tagged during this time may
become unshowable.